requirements.txt
liquipediapy2>2.0.0
nonebot
redis